+++ web/src/app/chat/message/SubQuestionsDisplay.tsx
@@ -9,7 +9,7 @@ import { FiSearch } from "react-icons/fi";
 import { OnyxDocument } from "@/lib/search/interfaces";
 import { BaseQuestionIdentifier, SubQuestionDetail } from "../interfaces";
 import { SourceChip2 } from "../input/ChatInputBar";
-import { ResultIcon } from "@/components/chat_search/sources/SourceCard";
+import { ResultIcon } from "@/components/chat/sources/SourceCard";
 import { openDocument } from "@/lib/search/utils";
 import { SourcesDisplay } from "./SourcesDisplay";
 import ReactMarkdown from "react-markdown";
