+++ web/src/app/chat/shared/[chatId]/SharedChatDisplay.tsx
@@ -17,11 +17,11 @@ import { OnyxInitializingLoader } from "@/components/OnyxInitializingLoader";
 import { Persona } from "@/app/admin/assistants/interfaces";
 import { Button } from "@/components/ui/button";
 import { OnyxDocument } from "@/lib/search/interfaces";
-import TextView from "@/components/chat_search/TextView";
+import TextView from "@/components/chat/TextView";
 import { DocumentResults } from "../../documentSidebar/DocumentResults";
 import { Modal } from "@/components/Modal";
-import FunctionalHeader from "@/components/chat_search/Header";
-import FixedLogo from "../../shared_chat_search/FixedLogo";
+import FunctionalHeader from "@/components/chat/Header";
+import FixedLogo from "../../../../components/logo/FixedLogo";
 import { useRouter } from "next/navigation";
 
 function BackToOnyxButton({
