+++ web/src/components/chat/BlurBackground.tsx
