+++ web/src/components/logo/FixedLogo.tsx
